# services/academic-core-service/docs/openapi-front-office.yaml
...
openapi: 3.0.3
info:
  title: Academic Core Service – Front Office API
  description: |
    API for managing front office operations within the academic-core-service.
    This module handles day-to-day administrative activities including:
    - Visitors Log
    - Phone Call Log
    - Half Day Notices
    - Postal Dispatch/Receive
    - Admission Enquiries
    - Complaints
    
    All operations are scoped to a specific school and academic year.
  version: 1.0.0
servers:
  - url: /api/v1
    description: Base API path
...
paths:
...
  /front-office/visitors:
    get:
      summary: List visitors
      description: |
        Retrieve a paginated list of visitors with optional filters.
        Supports filtering by purpose, date range, and search by name/phone.
      tags: [Visitors]
      parameters:
        - $ref: '#/components/parameters/SchoolIdHeader'
        - $ref: '#/components/parameters/AcademicYearIdHeader'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SortParam'
        - name: purpose
          in: query
          schema:
            type: string
          description: Filter by visit purpose
        - name: fromDate
          in: query
          schema:
            type: string
            format: date
          description: Filter visitors from this date
        - name: toDate
          in: query
          schema:
            type: string
            format: date
          description: Filter visitors until this date
        - name: search
          in: query
          schema:
            type: string
          description: Search by visitor name or phone number
        - name: checkedOut
          in: query
          schema:
            type: boolean
          description: Filter by checkout status (true = checked out, false = still in)
      responses:
        '200':
          description: Paginated list of visitors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitorPageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Add visitor
      description: Record a new visitor entry.
      tags: [Visitors]
      parameters:
        - $ref: '#/components/parameters/SchoolIdHeader'
        - $ref: '#/components/parameters/AcademicYearIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVisitorRequest'
      responses:
        '201':
          description: Visitor recorded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitorResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /front-office/visitors/{id}:
    get:
      summary: Get visitor details
      description: Retrieve details of a specific visitor entry.
      tags: [Visitors]
      parameters:
        - $ref: '#/components/parameters/SchoolIdHeader'
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Visitor details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitorResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /front-office/visitors/{id}/checkout:
    patch:
      summary: Update visitor checkout time
      description: Record the checkout time for a visitor.
      tags: [Visitors]
      parameters:
        - $ref: '#/components/parameters/SchoolIdHeader'
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VisitorCheckoutRequest'
      responses:
        '200':
          description: Checkout time updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitorResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

  # -------------------------------------------------------------------------
  # PHONE CALL LOG
  # -------------------------------------------------------------------------
...
components:
...
  parameters:
    SchoolIdHeader:
      name: X-School-Id
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: The school context for the request
    AcademicYearIdHeader:
      name: X-Academic-Year-Id
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: The academic year context for the request
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
      description: Page number (0-indexed)
    SizeParam:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Number of items per page
    SortParam:
      name: sort
      in: query
      schema:
        type: string
      description: Sort criteria (e.g., 'createdAt,desc')

  # ---------------------------------------------------------------------------
  # RESPONSES
  # ---------------------------------------------------------------------------
  responses:
    BadRequest:
      description: Bad request - validation errors
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflict - operation not allowed in current state
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  # ---------------------------------------------------------------------------
  # SCHEMAS
  # ---------------------------------------------------------------------------
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
    PageMetadata:
      type: object
      properties:
        page:
          type: integer
          description: Current page number (0-indexed)
        size:
          type: integer
          description: Number of items per page
        totalElements:
          type: integer
          format: int64
          description: Total number of elements
        totalPages:
          type: integer
          description: Total number of pages

    # --- Enums ---
    FrontOfficeStatusEnum:
      type: string
      enum: [ACTIVE, CLOSED, ARCHIVED]
      description: Status for front office records
...
    VisitorResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        visitorName:
          type: string
        phoneNumber:
          type: string
        purpose:
          type: string
        numberOfPersons:
          type: integer
        idProofType:
          type: string
          nullable: true
        idProofNumber:
          type: string
          nullable: true
        checkInTime:
          type: string
          format: date-time
        checkOutTime:
          type: string
          format: date-time
          nullable: true
        remarks:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/FrontOfficeStatusEnum'
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
          format: uuid
    CreateVisitorRequest:
      type: object
      required: [visitorName, phoneNumber, purpose, numberOfPersons, checkInTime]
      properties:
        visitorName:
          type: string
          minLength: 2
          maxLength: 100
        phoneNumber:
          type: string
          pattern: '^\+?[0-9]{10,15}$'
        purpose:
          type: string
          minLength: 2
          maxLength: 255
        numberOfPersons:
          type: integer
          minimum: 1
          default: 1
        idProofType:
          type: string
          maxLength: 50
        idProofNumber:
          type: string
          maxLength: 50
        checkInTime:
          type: string
          format: date-time
        remarks:
          type: string
          maxLength: 500
    VisitorCheckoutRequest:
      type: object
      required: [checkOutTime]
      properties:
        checkOutTime:
          type: string
          format: date-time
        remarks:
          type: string
          maxLength: 500
    VisitorPageResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/VisitorResponse'
        page:
          $ref: '#/components/schemas/PageMetadata'

    # --- Phone Calls ---
...